Created a crazy byte wrapper class
Just created a class that can use a byte array to create a bit arrays
of any size - numbers 1 bits - inf bits long, and can convert between
different sized bit arrays super easily. This will be incredibly useful
for using as little space as possible when storing chunk data in
memory. Not only that, but it’s backed by the dynamically compressing
byte array, so it compresses itself when it’s not in use!

Crazy stuff