Shifted responsibilities around a bit
Regions are now responsible for saving themselves, but it can still be
triggered from another object for example when disabling the plugin.
Made the MirrorWorld class a bit less cluttered.

Also removed the responsibilty of notifying when a room is added from
chunks, as it was complicating things and should probably just be
handled by one of the containing classes. Notifying regions when chunks
have changed has also been removed for now just because chunks are
already complicated enough