Added a default config.yml and let that decide the settings
It is now possible to specify in the config.yml whether the different
types of hiding should be enabled on a per-world basis. Settings that
were only specific to certain parts of the program have been moved into
that part of the program.

It is also possible to specify in the config.yml which types of
entities should be hidden with the entity hider.